requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
requests==2.32.3
twilio==9.2.3
python-dotenv==1.0.1
python-multipart==0.0.9
psycopg2-binary
chardet
